ai-core/hybrid-engine-v0.1/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hybrid-engine"
version = "0.1.0"
description = "Cockswain Hybrid Engine v0.1 (Echo core + Memory hook)"
authors = [{name="Cockswain", email="noreply@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.6.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.1",
  "prometheus-client>=0.20.0"
]

[project.scripts]
hybrid-engine = "hybrid_engine.__main__:main"